# packages/shared-schema/openapi.json
{
  "components": {
    "schemas": {
...
      "ActionResponsePayload": {
        "description": "Action execution response.",
        "properties": {
          "effects": {
            "items": {
              "$ref": "#/components/schemas/StackingEffectPayload"
            },
            "title": "Effects",
            "type": "array"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/DomainEventPayload"
            },
            "title": "Events",
            "type": "array"
          },
          "missile_outcome": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MissileOutcomePayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "morale_outcomes": {
            "items": {
              "$ref": "#/components/schemas/MoraleOutcomePayload"
            },
            "title": "Morale Outcomes",
            "type": "array"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "pending_tq_checks": {
            "items": {
              "$ref": "#/components/schemas/PendingTQCheckPayload"
            },
            "title": "Pending Tq Checks",
            "type": "array"
          },
          "pursuit_outcome": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PursuitOutcomePayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "shock_outcome": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ShockOutcomePayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "state": {
            "$ref": "#/components/schemas/GameStatePayload"
          },
          "tq_check_outcomes": {
            "items": {
              "$ref": "#/components/schemas/TQCheckOutcomePayload"
            },
            "title": "Tq Check Outcomes",
            "type": "array"
          }
        },
        "required": [
          "ok",
          "reason",
          "state",
          "effects",
          "pending_tq_checks",
          "tq_check_outcomes",
          "morale_outcomes",
          "events"
        ],
        "title": "ActionResponsePayload",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LegalMovesPayload": {
        "description": "Legal move options for one unit under current game state.",
        "properties": {
          "options": {
            "items": {
              "$ref": "#/components/schemas/LegalMoveOptionPayload"
            },
            "title": "Options",
            "type": "array"
          },
          "unit_id": {
            "title": "Unit Id",
            "type": "string"
          }
        },
        "required": [
          "unit_id",
          "options"
        ],
        "title": "LegalMovesPayload",
        "type": "object"
      },
      "LoadGamePayload": {
        "description": "Load snapshot options.",
        "properties": {
          "slot_id": {
            "default": "quicksave",
            "title": "Slot Id",
            "type": "string"
          }
        },
        "title": "LoadGamePayload",
        "type": "object"
      },
      "MissileActionPayload": {
        "description": "Missile command payload.",
        "properties": {
          "fire_mode": {
            "default": "active",
            "enum": [
              "active",
              "reaction"
            ],
            "title": "Fire Mode",
            "type": "string"
          },
          "firing_unit_id": {
            "title": "Firing Unit Id",
            "type": "string"
          },
          "modifier_ids": {
            "items": {
              "type": "string"
            },
            "title": "Modifier Ids",
            "type": "array"
          },
          "reaction_trigger": {
            "anyOf": [
              {
                "enum": [
                  "entry",
                  "retire",
                  "return"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reaction Trigger"
          },
          "target_unit_id": {
            "title": "Target Unit Id",
            "type": "string"
          }
        },
        "required": [
          "firing_unit_id",
          "target_unit_id"
        ],
        "title": "MissileActionPayload",
        "type": "object"
      },
...
      "MissileReloadActionPayload": {
        "description": "Missile reload command payload.",
        "properties": {
          "unit_id": {
            "title": "Unit Id",
            "type": "string"
          }
        },
        "required": [
          "unit_id"
        ],
        "title": "MissileReloadActionPayload",
        "type": "object"
      },
...
      "MoveActionPayload": {
        "description": "Move command payload.",
        "properties": {
          "destination": {
            "$ref": "#/components/schemas/HexPayload"
          },
          "unit_id": {
            "title": "Unit Id",
            "type": "string"
          }
        },
        "required": [
          "unit_id",
          "destination"
        ],
        "title": "MoveActionPayload",
        "type": "object"
      },
...
      "ShockActionPayload": {
        "description": "Shock combat command payload.",
        "properties": {
          "angle": {
            "default": "front",
            "enum": [
              "front",
              "flank",
              "rear"
            ],
            "title": "Angle",
            "type": "string"
          },
          "attacker_unit_id": {
            "title": "Attacker Unit Id",
            "type": "string"
          },
          "defender_unit_id": {
            "title": "Defender Unit Id",
            "type": "string"
          },
          "modifier_ids": {
            "items": {
              "type": "string"
            },
            "title": "Modifier Ids",
            "type": "array"
          }
        },
        "required": [
          "attacker_unit_id",
          "defender_unit_id"
        ],
        "title": "ShockActionPayload",
        "type": "object"
      },
...
      "SnapshotPayload": {
        "description": "Save/load response payload including state.",
        "properties": {
          "event_offset": {
            "title": "Event Offset",
            "type": "integer"
          },
          "saved_at": {
            "title": "Saved At",
            "type": "string"
          },
          "slot_id": {
            "title": "Slot Id",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/GameStatePayload"
          }
        },
        "required": [
          "slot_id",
          "saved_at",
          "event_offset",
          "state"
        ],
        "title": "SnapshotPayload",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Legions API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/game/action": {
      "post": {
        "description": "Apply movement action and return updated state payload.",
        "operationId": "game_action_game_action_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveActionPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponsePayload"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Game Action",
        "tags": [
          "game"
        ]
      }
    },
    "/game/action/missile": {
      "post": {
        "description": "Apply missile action and return updated state payload.",
        "operationId": "missile_action_game_action_missile_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MissileActionPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponsePayload"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Missile Action",
        "tags": [
          "game"
        ]
      }
    },
    "/game/action/missile/reload": {
      "post": {
        "description": "Apply missile reload action and return updated state payload.",
        "operationId": "missile_reload_action_game_action_missile_reload_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MissileReloadActionPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponsePayload"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Missile Reload Action",
        "tags": [
          "game"
        ]
      }
    },
    "/game/action/shock": {
      "post": {
        "description": "Apply shock action and return updated state payload.",
        "operationId": "shock_action_game_action_shock_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShockActionPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponsePayload"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Shock Action",
        "tags": [
          "game"
        ]
      }
    },
...
    "/game/legal-moves/{unit_id}": {
      "get": {
        "description": "Return legal move destinations and path preview metadata for one unit.",
        "operationId": "legal_moves_game_legal_moves__unit_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "unit_id",
            "required": true,
            "schema": {
              "title": "Unit Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LegalMovesPayload"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Legal Moves",
        "tags": [
          "game"
        ]
      }
    },
    "/game/load": {
      "post": {
        "description": "Load previously saved snapshot into active game state.",
        "operationId": "load_game_game_load_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoadGamePayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotPayload"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Load Game",
        "tags": [
          "game"
        ]
      }
    },
    "/game/new": {
      "post": {
        "description": "Reset in-memory game and return the fresh state.",
        "operationId": "new_game_game_new_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewGamePayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameStatePayload"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "New Game",
        "tags": [
          "game"
        ]
      }
    },
    "/game/preview/missile": {
      "post": {
        "description": "Return read-only missile preview details without mutating game state.",
        "operationId": "missile_preview_game_preview_missile_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MissileActionPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MissilePreviewResponsePayload"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Missile Preview",
        "tags": [
          "game"
        ]
      }
    },
    "/game/preview/shock": {
      "post": {
        "description": "Return read-only shock preview details without mutating game state.",
        "operationId": "shock_preview_game_preview_shock_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShockActionPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShockPreviewResponsePayload"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Shock Preview",
        "tags": [
          "game"
        ]
      }
    },
...
    "/game/save": {
      "post": {
        "description": "Persist current game state snapshot under selected slot id.",
        "operationId": "save_game_game_save_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveGamePayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotPayload"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Game",
        "tags": [
          "game"
        ]
      }
    },
...
  }
}